Extract and reuse UI methods to make clickable labels and browse urls - We will use that with the update notification link later
